chore(ideavimrc): add clipboard setting and additional key mappings feat(ideavimrc): add quick definition mappings and easy motion plugin settings The 'set clipboard=unnamed' setting is added to enable the use of the system clipboard in Vim. This allows for easier copying and pasting between Vim and other applications.
The 'nnoremap gh :action QuickJavaDoc<CR>' mapping is added to provide a quick way to view the documentation for a symbol under the cursor. Pressing 'gh' will trigger the 'QuickJavaDoc' action.

The 'noremap <F12> :action GotoImplementation<CR>' mapping is added to provide a quick way to navigate to the implementation of a symbol under the cursor. Pressing 'F12' will trigger the 'GotoImplementation' action.

The 'sethandler <c-p> a:vim' and 'noremap <c-p> :action GotoAction' lines