feat(lsp): add Docker Language Server support
feat(mason-lsp): add Docker Language Server to the list of supported language servers
feat(null-ls): add Hadolint to the list of supported linters
The Docker Language Server is now supported in the LSP configuration. It has been added to both the `lsp.lua` and `mason-lsp.lua` files. Additionally, Hadolint has been added to the list of supported linters in the `null-ls.lua` file. This allows for better Docker-related language support and linting capabilities in the editor.